Moved account API requests to the apis directory
