+++ src/js/containers/account/filters/AccountObjectClassContainer.jsx
@@ -11,7 +11,7 @@ import { isCancel } from 'axios';
 
 import * as accountFilterActions from 'redux/actions/account/accountFilterActions';
 
-import * as AccountHelper from 'helpers/accountHelper';
+import * as AccountHelper from 'apis/account';
 
 import ObjectClassFilter from 'components/account/filters/objectClass/ObjectClassFilter';
