+++ src/js/containers/account/filters/AccountProgramActivityContainer.jsx
@@ -11,7 +11,7 @@ import { connect } from 'react-redux';
 import { isCancel } from 'axios';
 
 import * as accountFilterActions from 'redux/actions/account/accountFilterActions';
-import * as AccountHelper from 'helpers/accountHelper';
+import * as AccountHelper from 'apis/account';
 
 import { _resetExchange, _convertToFrontendFilter } from 'models/v1/account/queries/queryBuilders/_programActivityTranslator';
