+++ tests/components/aboutTheData/TimeFilters-test.jsx
@@ -3,7 +3,7 @@ import { List } from 'immutable';
 import { render, screen } from '@test-utils';
 
 import TimeFilters from 'components/aboutTheData/TimeFilters';
-import * as accountHelpers from 'helpers/accountHelper';
+import * as accountHelpers from 'apis/account';
 import * as glossaryHelpers from 'helpers/glossaryHelper';
 import { mockGlossary } from '../../mockData';
 import { mockSubmissions } from '../../mockData/helpers/aboutTheDataHelper';
