+++ tests/containers/account/AccountContainer-test.jsx
@@ -12,7 +12,7 @@ import FederalAccount from 'models/v1/account/FederalAccount';
 
 import { mockAccount, mockReduxAccount, mockSnapshot } from './mockAccount';
 
-jest.mock('helpers/accountHelper', () => require('./accountHelper'));
+jest.mock('apis/account', () => require('./accountHelper'));
 jasmine.DEFAULT_TIMEOUT_INTERVAL = 10000;
 
 // spy on specific functions inside the component
